+++ catalyst_voices/packages/internal/catalyst_voices_view_models/lib/src/campaign/current_campaign_info_view_model.dart
@@ -7,8 +7,8 @@ import 'package:equatable/equatable.dart';
 class CurrentCampaignInfoViewModel extends Equatable {
   final String title;
   final String description;
-  final int allFunds;
-  final int totalAsk;
+  final double allFunds;
+  final double totalAsk;
   final List<CampaignTimelineViewModel> timeline;
   final ComparableRange<Coin> askRange;
